Updating analysisDialog after supporting non-integer path lengths in API

+++ gui/analysisDialog.cpp
@@ -369,7 +369,7 @@ void AnalysisDialog::calculateDistances() {
     mw->progressDialog->setLabelText("Finding shortest paths in component ...");
     // calculate the shortest path lengths within it
     //for(unsigned int t = 0; t< (*giant_comp).size(); t++) cerr << (*giant_comp)[t]->get_id() << endl;
-    vector< vector<int> > pathLengths = network->calculate_unweighted_distances(giant_comp);
+    vector< vector<double> > pathLengths = network->calculate_unweighted_distances(giant_comp);
     double diam = 0.0;
     double mean = 0.0;
     for (unsigned int i = 0; i<pathLengths.size(); i++) {
